New command line argument "--skip_hash"
Just add it to the beginning of the command (./extract.py --skip_hash)
and the hash will be skipped.

Fixed bug where if the old file that matches the new file does not exist
we get a crash. Now we attempt to extract the new file without the old file.

Added automatic file renaming. You can put your .img files in the "old"
directory and they will be renamed automatically, and back at the end of
the extraction. Additionally, the "output" directory will get .img appended
to their file names after the extraction.

Removed the need for the "LD_LIBRARY_PATH=./lib64/" part of the
command by setting it within the script.

Updated README.md to reflect all changes.